Fixed two problems for 64-bit version.

+++ src/libgtmatch/runsuffixerator.c
@@ -233,7 +233,7 @@ int parseargsandcallsuffixerator(int argc,const char *argv[],Env *env)
   int retval;
   bool haserr = false;
 
-  printf("# sizeof(Seqpos)=%u\n",sizeof(Seqpos) * CHAR_BIT);
+  printf("# sizeof(Seqpos)=%lu\n",(unsigned long) (sizeof(Seqpos) * CHAR_BIT));
   retval = suffixeratoroptions(&so,argc,argv,env);
   if (retval == 0)
   {
